Creating checkout functionality
This took some time because it's hard. There I make a functionality that increase and decrease the quantity of items them how many you increase it will decrease in database and we also got the JWT of these items those we checkout and these details will display on orders page of user the admin orders page will be another one where we got all over the orders and when we done we will delete the order as we need.